dependencies: Update jsonnet libraries
This change updates:

* go-jsonnet Golang library, to the latest master version
* base (C++) jsonnet library, to the latest released version

in order to pick up improvements to the jsonnet stdlib.

Dependencies to go-jsonnet are loaded via bazel macro, and this macro is
not well-behaved (attempts to reconfigure the Golang toolchain) so this
functionality is patched out.

go-jsonnet also tends to load an outdated version of the base C++
library, so an explicit load is added for this library at the latest
released version.

Tested: `bazel test //...` still passes