enkit bazel affected-targets: Remove presubmit skipping
This change removes the cursed bit of logic that skips the presubmit if
the beginning of the build graph was broken; this allows the presubmit
to pass for a change that rolls back build graph breakage, which used to
be a problem.

This is still ostensibly a problem; however, now we have a more educated
team and a mechanism to force the presubmits to pass, so this bit of
logic can be removed, which is causing its own version of pain.

Tested: no

Jira: INFRA-7604