postsubmit: Add support for debugging steps
This change adds a mechanism for adding user-friendly debugging steps
for determining why a postsubmit isn't behaving properly.

The first step added is a "large file printer", which will print out the
largest N files in a particular directory; this will be useful for
debugging INFRA-8302.

Tested: N/A

Jira: INFRA-8302