gee: reset instead of rebase when restoring a branch (#983)
gee: reset instead of rebase when restoring a branch

When gee creates a new branch, it checks to see if a branch with the same
name exists in origin, and if it exists, it offers to restore the branch
from origin.

old behavior: gee would rebase the origin/branch onto master, which could
have merge conflicts.

new behavior: gee resets the branch to be identical to origin/branch, letting
the user choose when to rebase.

Tested: I used this feature in my build-00 container to restore a branch that
I had previously created from my build-04 container.  Works.