device plugin: Add internal interfaces, some plugin impl
This PR implements some of the top-level plugin logic:

* A "main" is added that looks like the main in the Nomad-provided
  skeleton example
* `PluginInfo`, `ConfigSchema`, `SetConfig`, and `Fingerprint` are
  implemented with high-level logic
* Unit tests for `Fingerprint` are added

This PR also defines interfaces for `Reserver` (how the plugin
communicates reservations) and `Notifier` (how the plugin receives
notifications that license state has changed, including from jobs on the
same node as well as jobs from other nodes). An implementation is
stubbed out for a postgres-backed table, and mocks are used for the
plugin unit tests.

Tested: unit tests added