gee: handle closed PRs that are also drafts (#1004)
gee: handle closed PRs that are also drafts

User exposed this bug: if a PR is marked as a draft, and then closed, github
labels it as state=CLOSED and isDraft=True.  Gee's old logic confused this
with an open PR with isDraft=True.  This new logic treats state as primary
and isDraft as secondary.

Tested: manual testing

```
  gcd -m testing_gee
  vi gee
  gee commit -a -m "blah blah blah"
  gee pr_make  # a draft PR
  gh pr close
  gee pr_make  # works now.
```