gee: add "--text" option to lspr (#1009)
gee: add "--text" option to lspr

Sometimes I have a lot of PRs that I need reviews for.  To help my reviewers,
it's useful to generate a list of open PRs in text format, suitable for pasting
into a chat message or email.

This PR adds a "--text" option to "gee lspr" to generate this.  It further
enhances the lspr functionality by reporting presubmit check status to
the report.

Tested:

```
$ ./gee lspr --text
* #1009: gee: add "--text" option to lspr
  2024-01-02 REVIEW_REQUIRED Checks pending.
  https://github.com/enfabrica/internal/pull/1009

* #1005: a test draft PR
  2023-12-18 REVIEW_REQUIRED DRAFT Checks passed.
  https://github.com/enfabrica/internal/pull/1005

* #840: //proxy/ptunnel:go_default_test: tagged no-presubmit
  2023-02-01 REVIEW_REQUIRED Checks passed.
  https://github.com/enfabrica/internal/pull/840

* #378: gee: starting to rewrite in golang
  2021-12-23 REVIEW_REQUIRED CHECKS FAILED.
  https://github.com/enfabrica/internal/pull/378

* #358: gee: explicit paths for enkit
  2021-12-08 REVIEW_REQUIRED CHECKS FAILED.
  https://github.com/enfabrica/internal/pull/358

```