license manager: Delete unused code
`flextape` has replaced `manager`, but the latter is still present in
the source tree. This change removes it so that it no longer needs to be
maintained.

The astore upload of the `manager` client has been replaced with one of
`flextape_client`, to finish the migration from one to the other.

Tested: `bazel test //...` still passes as much as it is expected to