docker: Migrate image build to `oci_*` rules
This change migrates all images built in this repo to use `rules_oci`
rules instead of `rules_docker` rules. Thankfully, all images built in
this repo today used hermetic rules, so there is an easy migration
path.

For rules that used `container_image`, there is a simple translation to
`oci_image`.

For rules that used `go_image`, there are extra transforms:

* `pkg_tar` must be used to prep the layer data. All data was put into
  `/enfabrica/bin` to avoid conflicting with top-level dirs.
* `entrypoint` (and possibly `cmd`) must be set on the resulting image
  to maintain compatibility with `go_image`.

When translating `oci_push` rules, `latest` was chosen as a remote tag
to be uniform across the repo.

Though `gcr.io` is going to be deprecated soon, the images are not moved
to a separate repository at this time.

Tested:
* `bazel test //...` still works
* Pushed a few of the images; pulled them to make sure they ran as
  expected

Jira: INFRA-9240