Update rules_oci to support docker config (#1029)
This change updates the rules_oci version so that the `config` field can
be used in `oci_pull`. Besides updating the `rules_oci` external repo to
stay up-to-date, the `config` field may be required to fix the
permissions issue in postsubmits. Currently, some postsubmits are
failing due to a 401 permissions error pulling containers from `gcr.io`.
Setting the `config` field to point to a docker `config.json` might
solve the issue.

Tested:
- `bazel build --override_repository=enkit=$HOME/enkit
//infra/puppet:puppet_server_image` to test that container builds are
not broken with the newer version of `rules_oci`

JIRA: INFRA-9736