gee: improve rebasing of children of submitted PRs (#1022)
gee: improve rebasing of children of submitted PRs

In a chat group, someone suggested an improvement to the flow to follow when
rebasing the child branch of a branch that has been merged.  This PR captures
that improvement: to update the child branch with all of the commits of the
unsquashed parent branch, before using `rebase --onto` to move only the new
commits from the child branch on top of master.

Tested: manually testing this for a week or so in my own work.  I am my own
guinea pig.