website: Default to `python312` runtime
This change updates the Python appengine runtime default from 2.7 to
3.12, as the former is deprecated and results in an error on push.

Code is not actually used for these sites, but a runtime must be
selected. Therefore, the version should have no impact on the actual
functionality of the site.

Comments in the code imply that the later Python runtime might not be as
friendly for pushing static sites, but testing has not yet shown this to
be an issue.

Tested: `bazel run //bazel/website/testdata:testsite-push` successfully
pushes a site; https://enfabrica-documentation.wn.r.appspot.com/ was
updated

Jira: INFRA-9747