buildbuddy: Update proto copybara (#1035)
This change updates the buildbuddy proto copybara to copy more relevant
protos, so that we can keep the protos up-to-date with the
currently-running Buildbuddy version.

Now that Buildbuddy has been updated, the protos we're coding against
are outdated, and we're not able to view responses properly. When
importing the updated protos, it's evident that they've expanded their
proto definitions greatly, so we need to increase the number of protos
we're importing. This change imports the set of protos required by
`invocation.proto` transitively. A follow-on PR will perform the actual
import and BUILD file adjustments.

Jira: INFRA-9987