bes_publisher: Fix propagation of build tags
This change modifies how "build tags" are propagated as BES pubsub
message attributes.

Previously, passing `--build_metadata=build_tag:foo=bar` would add a
Pubsub attribute of `bt:foo` = `bar` on all BES pubsub messages for this
build.

The problem is that filtering on `bt:foo` is not possible due to the `:`
character.

This change replaces `:` with `__` which is a legal character to filter
on when creating Pubsub subscriptions, and won't get conflated with
single `_` characters in the key.

Tested: unit tests only

Jira: INFRA-10127