bazel/utils/merge_kwargs.bzl: fix latent bug in merging dicts or dicts (#1055)
Problem:
When you have a source dict like {"key": { ... another dict ...}} where
"key" does NOT exist in the destination dict, the inner dict is copied
as
is, with a simple `dict(source["key"])` pretty much (outer dict is NEW,
independent, inner DICT is shared between SOURCE and DEST).

This is BAD: if the INNER dict is later modified, the SOURCE dict is
also
modified. This was caught by @salil-enf in
https://github.com/enfabrica/internal/pull/37246/commits/af987146705c809679404778d36b1dccfa22d4cf,
as bazel complained about a READ ONLY / FROZEN dict being modified.

In this PR:
- make dicts always recursively copied
- update unit tests

Fun fact:
The unit test was actually catching the bug, I suspect when I originally
edited it I paid attention to the first dict and last dict, but not
the middle one - so did not notice that the second dict was also getting
changes from the first edit! Wrong!