linux: update bazel rules for kernel_module cross compiling arm64
Update the kernel_module build process to handle cross compiling for
arm64.  A number of things happened along the way:

- pass CPU architecture in KernelTreeInfo

The KernelTreeInfo, used for building out of tree kernel modules, is
by definition tied to a CPU architecture, so might as well pass the
CPU architecture to the build system that way.

Passing CPU architecture via KernelTreeInfo makes the `archs` argument
to the kernel_modules_rule() rule unnecessary and that never worked
anyway.  A KernelTreeInfo can only be used by a single CPU
architecture.  So this patch removes the `archs` argument.

- for arm64 create and use a bazel transition rule

When building for arm64, two rules are needed instead of only the one.
First the regular kernel_module_rule() is emitted, followed by a
transition rule that transitions bazel to an arm64 platform.  From
there toolchain resolution sorts out an appropriate arm64 toolchain to
use.

Signed-off-by: Curt Brune <curt@enfabrica.net>