Add support for cross compiling enf.ko for aarch64 (#1061)
This is now ready for review.

To actually build enf_core.ko for aarch64 requires the following PR from
the internal repo:
https://github.com/enfabrica/internal/pull/40341
Note: internal PR 40341 refers to kernel artifacts that were built by
this enkit PR.

Also I tested building a top of master internal enf_core.ko using this
enkit to make sure this enkit was backwardly compatible. It is. So
outside of review comments, this should be able to merge without
breaking the current internal master.

---------

Signed-off-by: Curt Brune <curt@enfabrica.net>