linux: skip non-host architecture kernel modules for CI
By default, bazel tries to build the non-host CPU architecture targets
for pre/post-submit, which fails because they don't match the host
architecture.  This is an old problem, see INFRA-1516.

This patch adds tags to those targets to skip them for
pre/post-submit.

The "transistioned" variants of the targets are built by
pre/post-submit so no coverage is lost.

Signed-off-by: Curt Brune <curt@enfabrica.net>