kbuild: add -c (clean) option to build-kernel-tree.sh
This patch adds a '-c' clean flag to the the build-kernel-tree.sh
script.  The default is not to clean, i.e. reuse the build directory.

This is convenient for development when iterating on the Linux kernel
and wanting incremental builds.

The kpub-astore.sh script calls this with -c set and will continue to
perform clean builds when generating astore artifacts.

Signed-off-by: Curt Brune <curt@enfabrica.net>