gee: Exit the add_parent_branches recursion when a branch has itself as a parent (#1075)
Using `gee` in a slightly non-standard way for the [fork-ns3
repo](https://github.com/enfabrica/fork-ns3). The `master` (which is
also the github default branch) is different from the branch that we
want to be doing our current development (`acf_s` branch). The way I am
trying to make this known to `gee` so that `gee rupdate` works is by
adding the entry:
`acf_s acf_s ''` in the `parents` file and changing
`_add_parent_branches_to_chain()` to make this a valid condition to end
the recursion.