Adds detection of "broken" tests: builds, but does not run them.
If a Bazel test rule is marked with the "broken" tag, then it
should not be run, as it's known to fail & pollutes test runs
with false negatives. However, if a broken test is ignored,
then it runs the danger of being forgotten about.

This change attempts to strike a balance between the two: broken
tests are not run, but they're A) required be associated with
a Jira ticket, and B) their names are printed during evaluation.