linux/defs.bzl: allow deps to be kernel-specific
Allow the interpolation of the kernel name into dependencies in the
`kernel_module` macro, so that users can depend on only the
kernel-specific dependency module for each built module rather than the
whole bundle.

With this change, `kernel_module` targets can do the following:

```
kernel_module(
    name = "foo",
    kernels = [
        "ubuntu-impish-generic",
        "ubuntu-impish-emu-aarch64",
    ],
    deps = [
        "//driver/bar:{kernel_name}-bar",
        "//driver/baz:{kernel_name}-baz",
        ":quux", # not all deps need to specify {kernel_name}
        ...
    ],
    ...
)
```

And the kernel name will be interpolated into `deps` at `{kernel_name}`
before deps is passed to the target being generated.

Background:

Currently when a kernel_module macro invocation uses multiple values in
`kernels`, there is the problem that other `kernel_module` invocations
which need to depend on one of the corresponding modules from the first
invocation must depend on the whole bundle to ensure that dependencies
for all of its modules are available.

This means that even if downstream targets properly depend on a specific
module from the second invocation, they will still indirectly depend on
all modules from the first invocation, i.e., you have a dependency
chain like the following:

//tests:foo_test  depends on
  //mods/mod1:ubuntu-impish-generic-mod1 depends on
      //mods/mod2:mod2

where `:ubuntu-impish-generic-mod1` properly selects a single module
selected for only the kernel that's actually used by the test, but
`:mod2` selects modules for *all* kernels, only one of which is actually
used by `:foo_test`.

This is currently needed because `kernel_module` has no way to express
"use this dependency for this kernel but that dependency for that
kernel"; this patch addresses that.