kernel_module - add paths to extra flags (#1094)
Currently, the Makefiles for kernel modules infer the bin directory from
the include flags, or set it to be `bazel-out/k8-fastbuild/bin`. Both
methods are brittle and the second one is incorrect if more than one
platform is targeted.

This change will allow kernel modules to point to the right directories
during compilation without having to resort to such brittle methods.