faketree: Fix exec_test failure outside dev container (#1139)
Prior to this change, `exec_test` was testing the version of `faketree`
baked into the dev container, rather than a version built from the
current source. Not only is this surprising, it doesn't work outside the
dev container.

This change adds the from-source `faketree` binary as a data dependency,
so that the test can find it using the `runfiles` library and run it
directly. Additionally, some of the usages of the `assert` library are
cleaned up to make the test easier to read.

Tested: test passes outside dev container