auth/server: start logging authentication requests and events.
Background:
Before this PR, the auth server had pretty much no logging. The only
debugging was possible by looking at nginx / appengine request logs,
that showed the URLs and gRPC endpoints hit by the client.

In this PR:
- log events through the authentication API. Specifically, attempts
  by user to obtain a token, start authentication, and so on.

The logging is done in a way that should be suitable to enable
log based metrics in GCP, one of the simplest way to introduce
metrics on appengine/cloudrun based projects.
(https://cloud.google.com/logging/docs/logs-based-metrics)

This is in preparation for the next PR, that will refactor the
logic in auth.go to make it more resilient to misbehaving users
and clients.