auth/server: drop jars once consumed.
Background:
The authentication process works like this:
1. The CLI client asks to start authentication. An URL to authenticate is returned.
2. The URL is hit by the user, using oauth/whatever. When the oauth authentication
   complete, the server has to remember a bunch of info to generate the actual
   authentication token. This is stored in a "jar".
3. The CLI client then retries the token. At this point, the "jar" no longer needs
   to be stored in memory.

Problem:
In step 3, the current code never deletes the JAR, it just remains in memory.
This was generally perceived to be ok: let's say each jar is 100 bytes
(it's smaller), it would take 1M authentication requests to consume 100MB of RAM,
and the backend has GBs of RAM, and is restarted when load is low.

In this PR:
As part of hardening the server, this PR removes the jar at the end of
step 3.

This does not fix the leak when for example the CLI does not retrieve the
token (no step 3 is run), or the user does oauth multiple times/accidentally
refreshes the token (multiple steps 2). Which is pretty much indistinguishable
from normal usage without maintaining more state.

Future work will need to add a periodic scrubber to remove unused jars.

Still... let's say a user refresh on same token happens once in 100
authentications, this PR makes it take 100 times longer to hit the
same limit.

Previous PR adding logging makes the issue also visible through
log based metrics.