auth/server (and client): change default retry and sleep parameters.
See comments for backgrounds and rationale.

In short: we want to move away from long lived http connections, which
are painful on appengine and similar platforms.

Short lived http connections means the client has to poll, retry frequently
if authentication does not complete in time.

This PR changes the defaults for the # of attempts a client performs
to complete authentication (from 5 to 1800), the minimum wait time
between attempts (from 10 seconds to 1), and the max duration of a
connection on the server side (from 30 minutes to 6).

This means that old clients will still be allowed a ~30 minutes to
login (6 mins * 5 = 30 minutes) - which was the original goal.

Once new clients are rolled out, we'll lower the 6 minutes further
to probably ~1-2 seconds - while still allowing 30 minutes or more
to complete authentication (1800 attempts * 1 second = ~30 minutes)