bazel: Patch rules_go to allow embedding generated files
This change adds a patch to `rules_go` that will be necessary for using
the `embedsrcs` attribute of `go_library` rules to embed generated
files; this in turn is necessary to remove uses of `go_embed_data` rules
in favor of `embedsrcs`.

The current `rules_go` version has a `go_embed_data` rule that is used
to embed files (source files or generated files) into a Go
library/binary. This is used by enkit to embed static web assets, as
well as some generated assets (e.g. machinist client embeds a built
shared library, that it copies to the host system when provisioning).

The `go_embed_data` rule is deprecated in favor of native Go embedding.
Outside of bazel, this is controlled by `//go:embed` directives in the
source files (see [embed package docs](https://pkg.go.dev/embed) for
more info). This rule is removed in future versions of rules_go, so use
of the rule in enkit is preventing rules_go upgrades (which in turn
blocks a migration to bzlmod and later versions of bazel).

The `embedsrcs` attribute does work for embedding static files, and
gazelle will even generate BUILD files correctly based on `//go:embed`
directives in the source for static files.

For generated files, there are two problems:
1. gazelle will not generate the necessary `embedsrcs` attribute values
1. even if targets are added to `embedsrcs` for generated files, the
   logic in rules_go won't allow the files to be found and used during
   the build.

This change fixes the second problem listed above by adding effectively
`bazel-bin` as an extra `-embedlookupdir` arg to the `rules_go` builder,
 which will allow generated files to be referenced by `//go:embed`
 directives by their full repo-relative path.

 This should be a relatively safe patch to make; adding extra lookup
 directories can only have a few possible negative effects:

 1. increased build time/resources due to path combinatorial explosion
    (unlikely given the current usage pattern)
 1. picking up the wrong file due to ambiguity (e.g. conflict between
    source and generated files with the same name/path) which is also
    currently unlikely

Tested: Allows portions of future PRs to build (TODO: list PRs here
after they are put up)