astore server: Embed configs using `embedsrcs`
This change modifies how client configs are embedded into the astore
server (if at all), removing uses of `go_embed_data` in favor of
`embedsrcs`.

An example config is added so that the directory is present, and to
prove that `gazelle` integration + building works as expected; however,
whether it is safe to include such a config by default in future builds
should be reviewed.

Tested: builds/unit tests only