Update Boringssl to 202502131457
Update boringssl to the lastest version used in master of GRPC. This change is needed to address compile errors that show up when compiling with optimizations. The comile error is fixed in this newer version. The newer version also fixes the patch 0001. Removing patch 0002 (there should be no need to remvoe the ARM code, given it is blocked by macros.