Local kernel development improvements (#1184)
A couple of improvements for local kernel development:
- Allows having multiple installs under `~/rootfs` by adding `{version}`
to glob.
- Allows overriding `version` with KVER environment variable.

Also includes a small adjustment to protect against cache poisoning. The
local
kernel build tree is now an undeclared input so requiring the user
passes an
absolute path for `--@enkit//:kernel_dir` ensures two users builds wont
collide
with each other. This is of course not bullet proof.