gee pr_checks: handle old and new build URL parsing (#1202)
The URL that gee parses to extract the build number recently changed.

This is a point fix that lets gee handle both the old and new URL
syntax, so that both old and new PRs will work.

Also, Rolf pointed out that "_error" isn't defined.  I could've sworn
it used to be there.  Adding it (back?).

also deleting a spurious version of gee that seems to have been dropped
into the wrong directory.

Tested: manually tested.