gee: fix runaway rebase operations (#1204)
gee: fix runaway rebase operations

A month ago, I introduced a bug into gee that caused an infinite loop while
doing a `gee update` operation.

The problem turned out to be a change to the `_is_rebase_in_progress` function
that I learned about on stackoverflow.  Turns out the author was wrong, and
that the function would sometimes return false positives.

I reverted to my earlier implementation of this function to fix the issue.

Tested:

I had a branch that reproduced the issue.  I used `git reset --hard branch.REBASE_BACKUP` to
go back to the initial state, and re-tried `gee up` with the patched gee.  The problem was
resolved.