bazel/init: explicit bring in google benchmark as a dependency.
Background:
In the last commit I added a library that depended on com_github_google_bench.
It built, it worked. Except CI/CD was failing on missing libpfm dependency.
On further inspection, it turns out that com_github_google_bench is not included
anywhere in our WORKSPACE file, or bzlmod equivalent file.

So we're getting google bench indirectly, via another dependency
(likely, grpc / protocol buffers are bringing it in).

In our internal/ repository, which uses enkit, we explicitly bring in
googlebench 1.5.1. Turns out that the rule there is ignored.
A simple build of a benchmark depending on google bench shows that we
are linking against version 1.9.0!!!

Which makes sense: enkit is imported first, in WORKSPACE file the
**first** definition wins. So the rule we have in our internal repository
is useless, and in enkit we are picking whichever random version is
brought in by 3rd party dependencies.

In this commit:
- explicitly bring in google bench 1.9.2
- invoke the macro that allows to bring in dependencies of google bench

Verified:
Built my library with google bench dependency. The version of google
bench shown is now 1.9.2 instead of 1.9.0.
Similar result when using --override_repository for enkit in our
internal repo.

After this PR is merged, I'll remove the redundant definitions.