codeowners: Define root owners (#1214)
This PR adds myself, carlo and jason as root owners of the enkit repo.
The CODEOWNERS seems to be acting a little odd on this repo, assigning
people who don't exist in this file as CODEOWNERS. I'm hoping that
kicking this file a bit here will get the CODEOWNERS working.