Detect cycles in affected target dependency graphs (#1215)
`bazel query` is resilient to circular dependencies, meaning it will
still run and not report them
(https://github.com/bazelbuild/bazel/issues/12204#issuecomment-1154017067).
This is unfortunate, because our "affected targets" code is not able to
handle a cycle and will blow up with a stack overflow.

Detect cycles when querying deps, and print the first dep that led to
the cycle being discovered. This should help debug the issue when it
occurs. Unit test included.

Context:
https://chat.google.com/room/AAAAuo7llA8/LVt5EN1qWXg/LVt5EN1qWXg?cls=10