astore: URL fetch with UID+path sets empty TagSet on Retrieve (#1229)
This change ensures that when a URL-based download is happening by UID
that an empty TagSet gets put on the `RetrieveRequest`, to avoid the
default behavior of a `latest` tag filter being applied. Some logging is
added on failures, so that if similar problems persist we might be able
to debug them more quickly.

Tested: unit tests added (fails on master, passes with this change)

Jira: REL-240