enkit: Add logging around `login` hang (#1241)
This change cleans up logging in `enkit login` to facilitate debugging
of the token fetch hang.

With this change, it becomes clear that the hang is due to a buildbarn
request via bb_clientd after the token is fetched and saved, and has
nothing to do with the auth server.

Tested: ran `bazel run //enkit -- login --loglevel-console=debug` and
saw expected output.

Jira: INFRA-11783