astore_upload: Make wrapper tool switchable on the commandline (#1253)
This change makes the wrapper tool for `astore_upload` rules switchable
on the command-line, which will be useful for release builds where we
want the tool to have extra behavior, and we don't want to expose this
behavior publicly.

This would look something like:

```
bazel run \
  --@enkit//f/astore:uploader=//my/custom/upload:tool \
  //my/astore/upload:target
```

Tested: manually; `astore_upload` rules have same behavior as before.