mcu: Add perf to arm kernel build
perf is a utility to record performance statistics, useful for optimizing
applications which run on linux. It is related to kernel version as it
requires access to counters etc.  As such, as part of the arm build make
perf as well, so it is available to the Mcud team to profile applications
on the E1s.

Tickets:
MCU-3984